User-Service/requirements.txt
fastapi>=0.104.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
PyJWT>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

